kernel32/tests: Change file name to avoid collision with other tests.
Entries in IniFileMapping are persisted in Windows until the next
reboot, even if removed from the registry. This made the
function test_registry_mapping to cause other tests
that accessed the actual winetest.ini file to fail if the
profile test was run more than once on Windows without rebooting.

Signed-off-by: Carlos Rivera <carlos@superkaos.org>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>