ole32/tests: Fix flaky test in cowait_unmarshal_thread.
When run with runtest -q (WINETEST_DEBUG=0) it passed within the
first 50ms, but with any verbose logging it sometimes failed.

Use CoWaitForMultipleHandles (which keeps pumping the STA) since
cowait_unmarshal_thread only releases its marshaled unk at the end,
so it needs to be able to call back in order to exit cleanly.

Signed-off-by: Kevin Puetz <PuetzKevinA@JohnDeere.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>