Revert "winegstreamer: Avoid performing color matrix conversions in videoconvert.".
This reverts commit fd25ba65e0eb9fedfb2cdfa2b7a4b16e0401dfdf.

The relevant commit breaks visual correctness in many applications, for the sake
of performance. As such I am rather inclined to revert it for 6.0, especially as
a full solution will probably involve a revert of this patch along with upstream
changes to GStreamer.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=49535
Signed-off-by: Zebediah Figura <z.figura12@gmail.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>