include: Check whether the ms_abi attribute is supported on aarch64.
While wine's configure checks and requires that the ms_abi attribute
is supported on aarch64 (without it, variadic functions don't have the
correct ABI), these headers are also included when building widl as a
generic cross compilation tool as part of mingw-w64-tools. In the case
of widl, the functions that use these attributes (and in particular,
their ABI) doesn't matter as they aren't used/called, they're just
included as a side effect of including the headers that widl actually
needs.

This fixes building the widl tool for aarch64 linux with GCC, even if
wine itself can't be built in that configuration.

Only windef.h is used/needed by widl, but update msvcrt/corecrt.h as
well to keep these definitions in sync.

Signed-off-by: Martin Storsjo <martin@martin.st>
Signed-off-by: Jacek Caban <jacek@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>