winebus.sys: Remove incorrect Report ID from the bus_sdl HID report descriptor.
According to "USB Device Class Definition for Human Interface Devices
Firmware Specification Version 1.11" Report ID 0 is reserved and should
not be used.

On top of that if Report ID is used anywhere in the descriptor the
reports are expected to be prepended with an ID and this is not the case
for bus_sdl devices.

It worked only because our current HID parser implementation treats the
Report ID 0 as there would be none.

Signed-off-by: Arkadiusz Hiler <ahiler@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>