server: Don't check the user data for NULL in async_terminate().
This semantically reverts 481517178ff7dd542966d7f11f3d61b3fc8a0776.

That commit was used to implement NtFlushBuffersFile, which at the time didn't
use a callback function. 9050b58f0769fca75c4cebed7be3cce51989814e changed it to
use irp_completion(), since the result of a blocking flush needed to be taken
from the IOSB.

As of 97afac469fbe012e22acc1f1045c88b1004a241f that's not true anymore, but on
the other hand it is theoretically possible for a device driver to touch the
Information member of the IOSB, and we don't seem to lose anything by making
all asyncs take a common path.

Since all asyncs pass user data and there's no clear reason for them not to,
let's get rid of a bit of extra code complexity that's no longer used.

Signed-off-by: Zebediah Figura <zfigura@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>